refactor(reports): Restructured imports
